Describe a protocol in the library, and let SutterDevice speak through it
Problem: the protocol description lived inside its own test file, so nothing
could use it. SutterDevice still spoke through the commands dict of
communication/commands.py, whose DataCommand describes the protocol, builds the
bytes, performs the I/O, and then keeps the reply on itself -- on a class
attribute shared by every instance of the driver. It needed a second struct
format to write a position ('<lllc') beside the one that reads it ('<lllx'), and
those two had drifted: one of them silently packed a null byte where the
instrument sends a carriage return. Nothing could say that an acknowledgement must
be b"\r", so the driver checked it by hand three times. And the debug port was a
nested class with a process_command full of branches, holding a second table of
prefixes that had to be kept in step with the driver by hand -- badly, as it
turned out: it accepted a lowercase 'm' and a request with no terminator, neither
of which an MP-285 would, and two tests were passing against bytes that could
never have worked on a bench.

Solution: communication/protocol.py. A Frame turns named values into bytes and
bytes back into named values, owning no port and storing nothing. A binary frame
states one struct format, since pack and unpack are exactly each other's inverse;
a text frame states both its notations, since neither can be derived from the
other without guessing. Constants are stated once and serve both directions,
written on the way out and required on the way in, which is what closes both of
the Sutter's old gaps at once.

PhysicalDevice gains protocol (None by default) and performTransaction, which
performs one command under the port's transactionLock, reading by readLength for
a frame and up to the terminator for a line, through the primitives of
CommunicationPort and nothing else. Unlike sendCommand it does not require the
device to be Ready, which is exactly why the old SutterDevice had to reach around
sendCommand during initialization.

ProtocolDebugPort then needs no code of its own at all: whatever a request carries
is remembered by name, whatever a reply carries is answered from that memory, and
recognize() picks the command out of the same description the driver sends with.
One clause had to be invented for it -- "sets", for a state change the bytes
cannot express, since HOME carries nothing and still moves the stage. It is the
only part of a description that talks about the instrument rather than the wire,
and a driver never reads it.

CommandDictionary.validate() refuses a description that is wrong about itself: two
notations describing different lines, a struct packing a different number of
values than it names, a command whose request another command answers to first.
It makes specimen values from the declared types, so a protocol file is checked
without a line of test data. SutterDevice's own description is validated on every
run of its test.
